+++ src/jvmMain/kotlin/Main.kt
@@ -5,6 +5,10 @@ import io.peekandpoke.klang.audio_fe.create
 import io.peekandpoke.klang.audio_fe.samples.SampleCatalogue
 import io.peekandpoke.klang.audio_fe.samples.Samples
 import io.peekandpoke.klang.strudel.graal.GraalStrudelCompiler
+import io.peekandpoke.klang.strudel.lang.note
+import io.peekandpoke.klang.strudel.lang.slow
+import io.peekandpoke.klang.strudel.lang.sound
+import io.peekandpoke.klang.strudel.lang.stack
 import io.peekandpoke.klang.strudel.makeStatic
 import io.peekandpoke.klang.strudel.strudelPlayer
 import kotlinx.coroutines.delay
@@ -54,9 +58,14 @@ private suspend fun helloStrudel() {
     )
 
     strudel.use { strudel ->
-        val pattern = TestPatterns.active
+//        val pattern = TestPatterns.active
+//        val pattern1 = strudel.compile(pattern).await()
 
-        val pattern1 = strudel.compile(pattern).await()
+        val pattern1 = stack(
+            note("c3", "e3", "g3", "b3").sound("tri", "supersaw").slow(2)
+
+//            sound("bd", "hh", "sd", "oh").slow(1),
+        )
 
         println("=======================================================================")
         println(